ISIS-2726: inlines ClockServiceDefault with ClockService
no need for subclassing; is just a wrapper around InteractionLayerTracker service
